+++ src/core/page-tags/areas-mapping.js
@@ -1,0 +1,5 @@
+module.exports = {
+  'head': 'headTags',
+  'body-start': 'bodyStartTags',
+  'body-end': 'bodyEndTags'
+};